Moved parsePathData method from Graphics to new SVGPath module
